Add migration code for pem_base64 key
